test_message_send: Move out scheduled message tests.
